+++ docs/changelog.rst
@@ -47,6 +47,7 @@ New features
 
 * Added new attribute :py:attr:`botogram.Message.pinned_message`
 * Added new attribute :py:attr:`botogram.Sticker.emoji`
+* Added new method :py:meth:`botogram.Chat.leave`
 * Every method which sends something to a chat now returns the sent
   :py:class:`~botogram.Message`
 * Multiple instances of the same bot are now properly handled (as errors)